Allow for the sending of nil payloads.
If a blank string is used in place of nil to mean "no request entity",
Net::HTTP::HTTPGenericRequest#supply_default_content_type is called, which
sets the request's Content-Type to application/x-www-form-urlencoded.

This is problematic, because many servers will return a 406 Not Acceptable,
indicating that the given resource cannot accept request entities of that
type.

By using nil as the "no request entity" sigil, Net::HTTP::HTTPGenericRequest
will not automatically append a Content-Type to the request, resovling this
issue.