Add keep-alive support
For ease of implementation it rely on the excellent net-http-persistent.

Persistent connections are now the default, but can be turned off globaly with
```ruby
RestClient.persistent = false
```
Or on a request basis by passing `:persistent => false`

If you are not confident with making persistent connections the default you can
easily revert it by removing the `@persistent = true` line in `restclient.rb`.
The test suite will pass whatever the default is.