Store redirection history on reqs/responses.
This allows callers to actually determine whether a given request
followed redirection, and if so, what the original responses to each
intermediate request were.

Due to the way the code is organized, the simplest way to achieve this
involves storing the history on request objects as well. This isn't
ideal, but will work for now.