Use cookie jar for request cookies
Any cookies passed through a `:cookies` hash will be validated and
stored in a cookie jar for the request URL. Cookies passed in the
headers or payload will not be validated or stored, as before. The
cookie value can contain additional attributes, as if it was from a
Set-Cookie header.

This allows the cookies from a request to follow redirects if they match
the new URL.