Switch to using flatten_params everywhere.
Remove the old standalone implementation of Utils.encode_query_string in
favor of the unified version based on Utils.flatten_params.

This means that params in GET, POST x-www-form-urlencoded, and
POST multipart/form-data will all be handled in the same way (except for
escaping, since multipart/form-data is not escaped).