Allow IO-like objects for multipart payloads
The input object doesn't have to be an actual file, it can be just an
IO-like object. This is useful in projects like
[Refile](https://github.com/refile/refile) and
[Shrine](https://github.com/janko-m/shrine), which don't accept only
files but any IO object.

Internally it only needs to respond to `#read(*args)`, but I've made the
`FakeIO` test object with the 5 methods because we may need them in the
future (these are enough for object to be IO-like, we've determined that
in Refile and Shrine), so it's good to have a bit wider contract than
just `#read`.

I've needed this when building a shrine-cloudinary storage for Shrine,
because the Cloudinary gem internally uses rest-client.