Actually allow top level params to be ParamsArray.
Because of some idiosyncratic behavior in Request#process_url_params, we
would previously only handle params that were Hash objects. This is
really confusing behavior that stems from mixing options and headers.
Unfortunately we won't be able to clean that up until rest-client 3.0,
but at least allow ParamsArray or Hash for now.