Implement cookie jar support for :cookies.
The `:cookies` option may now be:
- A hash of strings
- An array of HTTP::Cookie objects
- A full HTTP::CookieJar

Rewrite all of the cookie processing on RestClient::Request to operate
in terms of HTTP::CookieJar, and expose `Request#cookie_jar` and
`Request#make_cookie_header` methods.

Raise ArgumentError when `:cookies` is provided in both request headers
and request options. Warn when both `:cookies` and a `Cookie` header are
provided.

Fix up tests for the new functionality as well.