Only do (expensive) SSL context setups when actually accessing HTTPS URLs
While upgrading our stack from CentOS 6.x to CentOS 7.x we noticed that between
rest-client 1.6.8 and 1.8.0 SSL certificate verification was added (yay! \o/),
which comes at some expense of SSL CA path setups and the like.

This expense cost us dearly, as we were doing a lot of backend requests via
CouchRest, which itself uses rest-client. For performance reasons, backend
communication is done unencrypted inside our secured network, so the SSL setup
was unnecessary. Our application code still needs to be ported to newer
versions of CouchRest and rest-client, so a batch for the 1.8.x branch was
needed.

This patch ensures SSL context setup is only done when accessing https URLs,
and updates related tests. Introduction of security issues by this PR is very
unlikely.

Forward-porting of this patch is recommended, additional recommended changes
might be caching/pooling of request objects or pre-creation/sharing of SSL CA
path information. The possible security implications of this additional change
where not explored in the scope of the current PR.