Major breaking change: switch to keyword args.
Use keyword arguments for RestClient::Request#initialize. This means
that when you call RestClient::Request.new() it will automatically
require :url and :method. It will also raise ArgumentError for
unexpected options.

Change the API for the RestClient.get() helper methods so that they
accept a URL, possibly a body, and then *options* rather than headers.
This is the single biggest wart in the API that has been rankling for
ages.

Drop compatibility for all of the super hacky usage where options were
hidden inside the headers hash. No longer accept passing :params or
:cookies as part of the headers hash.

Also remove a few other bits of vestigial code, and remove the
not-very-useful SSLCertificateNotVerified exception.