ability to have a seperate logger per thread
This should allow users of library to do stuff like

```ruby
def log_request logger, &block
  RestClient.local_logger = logger
  yield
ensure
  RestClient.local_logger = nil
end
```

which can help to log desired requests without having a global state.
Without this, `RestClient.log` has to be set which means all the
requests in all threads will be logged. Even if we are not interested in
logging them.