Track start/end times on response objects.
Record the start and end times of a request (from the time just before
the request is initiated to the time that the response object is
created). Log the duration when we log the end of the request.

Also refactor the handling of `:raw_response => true` / streaming
requests to be a little less awkward, and create a uniform `#size`
method regardless of whether the response is a Response or RawResponse.

Fixes: #126