Skip netrc with Authorization header, up/downcase.
The fix for the Netrc user/password overriding a user-set Authorization
header that was committed in #374 was case-sensitive and does not cover
users who pass `:authorization` or `"authorization"` as the header name.

Add @processed_headers_lowercase so that case-insensitive tests like
this are easier in the future. It would be better if
`Request#stringify_headers` normalized both Symbols and Strings.
https://github.com/rest-client/rest-client/issues/599

Also add a test for the case-insensitive behavior.

Fixes: #524, #550