When creating a Tempfile for a raw response, preserve the file extension
When passing `raw_response: true` into `RestClient::Request.
execute`, a `RestClient::RawResponse` is returned (instead of a
`Response`) which provides access to a `Tempfile` containing
the response body via its `#file` method. This is useful if,
for example, you expect a binary response from the server.

At the moment, the `Tempfile` is created with a generic name
(`rest-client.` followed by some random characters introduced by
`Tempfile` itself). But this is quite annoying as it can be useful
particularly to hold on to the original file extension (for
example, even `rest-client` [itself][1] uses the path of a
provided file to determine its `Content-Type` when performing
a file upload).

This tweaks the behaviour so that the `Tempfile` is created
with the same extension as was in the original URL.

[1]: https://github.com/rest-client/rest-client/blob/master/lib/restclient/payload.rb#L186